I'm going to push this, but then I'm going to branch and try something different.
Here's what I think needs to happen:  text needs to be stored raw, which it is.  Any custom markdown things can happen on update and save.

In particular, internal links should be changed into regular links to the pages they are connected to.

However, when a note is deleted, references to it will need to be deleted from all other notes as well, turning the text from a link back into regular text.